# src/server/routes/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "version": "1.0.0",
        "title": "UTD Degree Planner and Auditor",
        "description": "API Documentation for UTD Degree Planner and Auditor"
    },
    "servers": [
        {
            "url": "http://localhost:3000/"
        }
    ],
...
    "paths": {
...
        "/api/buildCourse/course": {
            "post": {
                "description": "Create a new course",
                "summary": "Create a new course",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Create a new course"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update a course",
                "summary": "Update a course",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Update a course"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/course/course": {
            "get": {
                "description": "Get a course by prefix and number",
                "summary": "Get a course by prefix and number",
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "name": "prefix",
                        "in": "path"
                    },
                    {
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "name": "number",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}